pyproject.toml
[project]
name = "watchdogs-security-city"
version = "0.1.0"
description = "OSINT Video/Image Analysis System with Multi-Agent AI Architecture"
authors = [
    { name = "WatchDogs Team", email = "team@watchdogs.dev" }
]
requires-python = ">=3.12"
dependencies = [
    "flask==3.0.0",
    "flask-cors==4.0.0",
    "flask-limiter==3.5.0",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.0",
    "langgraph>=0.2.50",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "openai>=1.58.1",
    "pillow>=11.0.0",
    "opencv-python-headless>=4.10.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "pydantic>=2.0.0",
    "folium>=0.15.0",
    "geopy>=2.4.0",
    "ffmpeg-python>=0.2.0",
    "piexif>=1.1.3",
    "reportlab>=4.0.0",
    "matplotlib>=3.8.0",
    "imagehash>=4.3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.23.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
    "pre-commit>=3.6.0",
    "types-pillow",
    "types-requests",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# ============================================================================
# MYPY Configuration (Strict Type Checking)
# ============================================================================
[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true
show_column_numbers = true

# Exclude files
exclude = [
    "^tests/",
    "^build/",
    "^dist/",
]

# Per-module options (for gradual typing)
[[tool.mypy.overrides]]
module = [
    "langchain.*",
    "langchain_openai.*",
    "langgraph.*",
    "folium.*",
    "geopy.*",
    "reportlab.*",
    "imagehash.*",
    "piexif.*",
    "flask_limiter.*",
]
ignore_missing_imports = true

# ============================================================================
# RUFF Configuration (Linting & Formatting)
# ============================================================================
[tool.ruff]
target-version = "py312"
line-length = 100
indent-width = 4

# Exclude directories
exclude = [
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    "venv",
    "__pycache__",
    "build",
    "dist",
    "*.egg-info",
]

[tool.ruff.lint]
# Enable rule categories
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "A",      # flake8-builtins
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimez
    "T10",    # flake8-debugger
    "ISC",    # flake8-implicit-str-concat
    "ICN",    # flake8-import-conventions
    "PIE",    # flake8-pie
    "PT",     # flake8-pytest-style
    "Q",      # flake8-quotes
    "RSE",    # flake8-raise
    "RET",    # flake8-return
    "SIM",    # flake8-simplify
    "PTH",    # flake8-use-pathlib
    "PL",     # pylint
    "RUF",    # ruff-specific rules
]

# Disable specific rules
ignore = [
    "E501",    # line-too-long (handled by formatter)
    "PLR0913", # too-many-arguments
    "PLR2004", # magic-value-comparison
    "ISC001",  # single-line-implicit-string-concatenation (conflicts with formatter)
]

# Allow autofix for all enabled rules
fixable = ["ALL"]
unfixable = []

# Allow unused variables when prefixed with underscore
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "PLR2004",  # magic values OK in tests
    "S101",     # assert OK in tests
    "ARG",      # unused args OK in tests (fixtures)
]
"__init__.py" = [
    "F401",     # unused imports OK in __init__
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# ============================================================================
# PYTEST Configuration
# ============================================================================
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-fail-under=70",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# ============================================================================
# COVERAGE Configuration
# ============================================================================
[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
    "*/.venv/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

# ============================================================================
# HATCH Build Configuration
# ============================================================================
[tool.hatch.build.targets.wheel]
packages = ["src"]
